add support for mobile rendering tests via `go test -tags mobile`
