+++ internal/driver/glfw/canvas_test.go
@@ -1,4 +1,5 @@
 // +build !ci
+// +build !mobile
 
 package glfw
