+++ internal/driver/glfw/menu_darwin_test.go
@@ -1,5 +1,6 @@
 // +build !ci
 // +build !no_native_menus
+// +build !mobile
 
 package glfw
