+++ internal/driver/glfw/menu_other_test.go
@@ -1,5 +1,6 @@
 // +build !ci
 // +build !darwin no_native_menus
+// +build !mobile
 
 package glfw
