+++ internal/driver/glfw/scale_test.go
@@ -1,4 +1,6 @@
-// +build !ci,!windows
+// +build !ci
+// +build !mobile
+// +build !windows
 
 package glfw
