+++ test/device_other.go
@@ -1,0 +1,7 @@
+// +build !mobile
+
+package test
+
+func (d *device) IsMobile() bool {
+	return false
+}
